Improve handling of post-load attribute changes
This commit adds a MutationObserver to listen for attribute changes
on <a> elements in case there are changes after page load. Right now
the observer won't catch all DOM changes that happen (perhaps from a bug
resulting from my not understanding the order in which page/content
scripts execute), so for now the workaround is we always wait for 1
second before querying for the links to instrument.